akasanoma-server/requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
falcon==2.0.0
gunicorn==19.9.0
importlib-metadata==0.23
more-itertools==7.2.0
numpy==1.17.2
packaging==19.2
pandas==0.25.1
peewee==3.11.2
pluggy==0.13.0
psycopg2-binary==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.2.1
python-datauri==0.2.8
python-dateutil==2.8.0
python-dotenv==0.10.3
python-http-client==3.2.1
pytz==2019.3
sendgrid==6.1.0
six==1.12.0
wcwidth==0.1.7
zipp==0.6.0
